Tests package updated
