+++ app/src/test/java/mangomessenger/tests/apis/sessions/DeleteCurrentSessionTests.kt
@@ -1,4 +1,4 @@
-package tests.apis.sessions
+package mangomessenger.tests.apis.sessions
 
 import mangomessenger.core.apis.SessionsApi
 import mangomessenger.core.apis.SessionsApiImpl
@@ -10,11 +10,11 @@ import mangomessenger.http.HttpInterceptor
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
-import tests.apis.variables.Credentials
+import mangomessenger.tests.apis.variables.Credentials
 import java.util.*
 import kotlin.collections.ArrayList
 
-class RefreshSessionTests {
+class DeleteCurrentSessionTests {
     private var loginResponse: LoginResponse = LoginResponse(null)
     private lateinit var httpClient: HttpClient
     private lateinit var sessionsApi: SessionsApi
@@ -28,18 +28,18 @@ class RefreshSessionTests {
     }
 
     @Test
-    fun refreshSessionSuccess() {
+    fun deleteCurrentSessionSuccess() {
         val loginRequest = LoginRequest(Credentials.Email, Credentials.Password)
         loginResponse = sessionsApi.loginAsync(loginRequest).get()
         val refreshToken = loginResponse.tokens?.refreshToken.toString()
-        val response = sessionsApi.refreshSessionAsync(refreshToken).get()
+        val response = sessionsApi.deleteCurrentSessionAsync(refreshToken).get()
         Assert.assertTrue(response.success)
     }
 
     @Test
-    fun refreshSessionUnauthenticated() {
+    fun deleteCurrentSessionUnauthenticated() {
         val refreshToken = UUID.randomUUID().toString()
-        val response = sessionsApi.refreshSessionAsync(refreshToken).get()
+        val response = sessionsApi.deleteCurrentSessionAsync(refreshToken).get()
         Assert.assertFalse(response.success)
     }
 }
